Additional template instantiations for int32_t==long

+++ modules/imgproc/src/fixedpoint.inl.hpp
@@ -171,9 +171,9 @@ class fixedpoint32
 private:
     int32_t val;
     fixedpoint32(int32_t _val) : val(_val) {}
-    static CV_ALWAYS_INLINE uint32_t fixedround(const uint32_t& _val) { return (_val + ((1 << fixedShift) >> 1)); }
+    static CV_ALWAYS_INLINE uint32_t fixedround(const uint32_t& _val) { return (_val + (((uint32_t)1 << fixedShift) >> (uint32_t)1)); }
 public:
-    static const int fixedShift = 16;
+    static const int32_t fixedShift = 16;
 
     typedef fixedpoint64 WT;
     typedef int32_t raw_t;
@@ -217,7 +217,7 @@ public:
     CV_ALWAYS_INLINE operator int32_t() const { return saturate_cast<int32_t>(); }
     CV_ALWAYS_INLINE bool isZero() { return val == 0; }
     static CV_ALWAYS_INLINE fixedpoint32 zero() { return fixedpoint32(); }
-    static CV_ALWAYS_INLINE fixedpoint32 one() { return fixedpoint32((1 << fixedShift)); }
+    static CV_ALWAYS_INLINE fixedpoint32 one() { return fixedpoint32(((int32_t)1 << fixedShift)); }
     friend class fixedpoint16;
 };
 
@@ -228,7 +228,7 @@ private:
     ufixedpoint32(uint32_t _val) : val(_val) {}
     static CV_ALWAYS_INLINE uint32_t fixedround(const uint32_t& _val) { return (_val + ((1 << fixedShift) >> 1)); }
 public:
-    static const int fixedShift = 16;
+    static const int16_t fixedShift = 16;
 
     typedef ufixedpoint64 WT;
     typedef uint32_t raw_t;
@@ -268,7 +268,7 @@ public:
     CV_ALWAYS_INLINE operator int32_t() const { return saturate_cast<int32_t>(); }
     CV_ALWAYS_INLINE bool isZero() { return val == 0; }
     static CV_ALWAYS_INLINE ufixedpoint32 zero() { return ufixedpoint32(); }
-    static CV_ALWAYS_INLINE ufixedpoint32 one() { return ufixedpoint32((1U << fixedShift)); }
+    static CV_ALWAYS_INLINE ufixedpoint32 one() { return ufixedpoint32(((uint32_t)1 << fixedShift)); }
 
     static CV_ALWAYS_INLINE ufixedpoint32 fromRaw(uint32_t v) { return ufixedpoint32(v); }
     CV_ALWAYS_INLINE uint32_t raw() { return val; }
@@ -293,7 +293,7 @@ public:
     CV_ALWAYS_INLINE fixedpoint16& operator = (const int8_t& _val) { val = ((int16_t)_val) << fixedShift; return *this; }
     CV_ALWAYS_INLINE fixedpoint16& operator = (const cv::softdouble& _val) { val = (int16_t)cvRound(_val * cv::softdouble((1 << fixedShift))); return *this; }
     CV_ALWAYS_INLINE fixedpoint16& operator = (const fixedpoint16& _val) { val = _val.val; return *this; }
-    CV_ALWAYS_INLINE fixedpoint16 operator * (const int8_t& val2) const { return cv::saturate_cast<int16_t>((int32_t)val * val2); }
+    CV_ALWAYS_INLINE fixedpoint16 operator * (const int8_t& val2) const { return cv::saturate_cast<int16_t>((int)val * val2); }
     CV_ALWAYS_INLINE fixedpoint32 operator * (const fixedpoint16& val2) const { return (int32_t)val * (int32_t)(val2.val); }
     CV_ALWAYS_INLINE fixedpoint16 operator + (const fixedpoint16& val2) const
     {
@@ -341,7 +341,7 @@ public:
     CV_ALWAYS_INLINE ufixedpoint16& operator = (const uint8_t& _val) { val = ((uint16_t)_val) << fixedShift; return *this; }
     CV_ALWAYS_INLINE ufixedpoint16& operator = (const cv::softdouble& _val) { val = _val.getSign() ? 0 : (uint16_t)cvRound(_val * cv::softdouble((int32_t)(1 << fixedShift))); return *this; }
     CV_ALWAYS_INLINE ufixedpoint16& operator = (const ufixedpoint16& _val) { val = _val.val; return *this; }
-    CV_ALWAYS_INLINE ufixedpoint16 operator * (const uint8_t& val2) const { return cv::saturate_cast<uint16_t>((uint32_t)val * val2); }
+    CV_ALWAYS_INLINE ufixedpoint16 operator * (const uint8_t& val2) const { return cv::saturate_cast<uint16_t>((int32_t)val * val2); }
     CV_ALWAYS_INLINE ufixedpoint32 operator * (const ufixedpoint16& val2) const { return ((uint32_t)val * (uint32_t)(val2.val)); }
     CV_ALWAYS_INLINE ufixedpoint16 operator + (const ufixedpoint16& val2) const
     {
